Move collation lookup key functions to shard file
